Added option to save egocentrically aligned videos. Added Docstrings.

+++ vame/util/align_egocentrical.py
@@ -94,7 +94,22 @@ def interpol(arr):
 
 def background(path_to_file,filename,video_format='.mp4',num_frames=1000):
     """
-    Compute background image from fixed camera 
+    Compute background image from fixed camera.
+    
+    Parameters
+    ----------
+    path_to_file : string
+        Project path from config.yaml.
+    filename : string
+        Name of video file to process (without extension).
+    video_format : string, optional (default '.mp4')
+        Format of video.
+    num_frames : int, optional (default 1000)
+        Number of frames to use for computing background.
+        
+    Returns
+    -------
+    Array of background values for video.
     """
     import scipy.ndimage
     capture = cv.VideoCapture(os.path.join(path_to_file,'videos',filename+video_format))
@@ -127,20 +142,41 @@ def background(path_to_file,filename,video_format='.mp4',num_frames=1000):
 
 def align_mouse(path_to_file,filename,video_format,crop_size, pose_list, 
                 pose_ref_index, confidence, pose_flip_ref,bg,frame_count,use_video=True):  
-    
-    #returns: list of cropped images (if video is used) and list of cropped DLC points
-    #
-    #parameters:
-    #path_to_file: directory
-    #filename: name of video file without format
-    #video_format: format of video file
-    #crop_size: tuple of x and y crop size
-    #dlc_list: list of arrays containg corresponding x and y DLC values
-    #dlc_ref_index: indices of 2 lists in dlc_list to align mouse along
-    #dlc_flip_ref: indices of 2 lists in dlc_list to flip mouse if flip was false
-    #bg: background image to subtract
-    #frame_count: number of frames to align
-    #use_video: boolean if video should be cropped or DLC points only
+    """
+    Parameters
+    ----------
+    path_to_file : string
+        Project path from config.yaml.
+    filename : string
+        Name of video to be processed (without extension)
+    video_format : string
+        Extension of filename.
+    crop_size : tuple of ints
+        Size of aligned data to return. Should be larger than the maximum single-frame motion.
+    pose_list : list
+        List of x,y coordinates & likelihoods for each body part in DLC format.
+    pose_ref_index : list
+        Reference bodyparts to perform egocentric alignment to.
+    confidence : float
+        Likelihood cutoff below which x,y values will be interpolated.
+    pose_flip_ref : list
+        indices of 2 lists in dlc_list to flip mouse if flip was false #TODO Clarify with creators.
+    frame_count : int
+        Number of frames in video
+    use_video : bool, optional (default True)
+        Whether to use the video for alignment or only pose data.
+
+
+    Returns
+    -------
+    images : list
+        if use_video=True, list of cropped and aligned images
+    points : list
+        List of egocentrically aligned poses.
+    time_series : list
+        List of frame indices for points and images.
+
+    """
     
     images = []
     points = []
@@ -223,7 +259,30 @@ def align_mouse(path_to_file,filename,video_format,crop_size, pose_list,
 
 
 #play aligned video
-def play_aligned_video(a, n, frame_count, path_to_file, filename=None, crop_size=None, save=False):
+def play_aligned_video(a, n, frame_count, path_to_file, save=False, filename=None, crop_size=None):
+    """
+    Parameters
+    ----------
+    a : array
+        Array or list of frames
+    n : list-like
+        Body parts for egocentric alignment
+    frame_count : int
+        Number of frames.
+    path_to_file : string
+        Project path from config.yaml.
+    save : bool, optional (default False)
+        Whether to save egocentrically aligned video.
+    filename : string, optional (default None)
+        Name of video to save if save=True.
+    crop_size : tuple, optional
+        Size of aligned video. Should be larger than the maximum single-frame motion.
+
+    Returns
+    -------
+    None.
+
+    """
     colors = [(255,0,0),(0,255,0),(0,0,255),(255,255,0),(255,0,255),(0,255,255),(0,0,0),(255,255,255)]
     if not os.path.exists(os.path.join(path_to_file, 'egocentricVideos/')):
         os.mkdir(os.path.join(path_to_file, 'egocentricVideos/'))
@@ -263,12 +322,41 @@ def play_aligned_video(a, n, frame_count, path_to_file, filename=None, crop_size
     cv.destroyAllWindows()
 
 
-def alignment(path_to_file, filename, pose_ref_index, video_format, crop_size, confidence, use_video=False, check_video=False, save=False):
-    
+def alignment(config, filename, pose_ref_index, video_format, crop_size, use_video=False, check_video=False, save=False):
+    """
+    Parameters
+    ----------
+    config : string
+        Path to config file
+    filename : string
+        Name of video to be processed (without extension)
+    pose_ref_index : list
+        List of 2 body parts to egocentrically align to.
+    video_format : string
+        Extension of filename.
+    crop_size : tuple
+        Size to crop aligned data to. Should be larger than maximum single-frmae movement.
+    use_video : Tbool, optional (default False)
+        Whether to use the video or only DLC poses.
+    check_video : bool, optional (Default False)
+        Whether to display the video. Press Q to exit while displaying.
+    save : bool, optional, (default False)
+        Whether to save the video instead of displaying.
+
+    Returns
+    -------
+    time_series : list
+        Frame indices
+    frames : list of arrays
+        Aligned poses or video frames.
+    """
+    config_file = Path(config).resolve()
+    cfg = read_config(config_file)
+    path_to_file = cfg['project_path']
+    confidence = cfg['pose_confidence']
     #read out data
-    data = pd.read_csv(os.path.join(path_to_file,'videos','pose_estimation',filename+'.csv'), skiprows = 2)
+    data = pd.read_csv(os.path.join(path_to_file,'videos','pose_estimation',filename+'.csv'), skiprows = 0, header=[1,2], index_col=0)
     data_mat = pd.DataFrame.to_numpy(data)
-    data_mat = data_mat[:,1:] 
     
     # get the coordinates for alignment from data table
     pose_list = []
@@ -300,12 +388,12 @@ def alignment(path_to_file, filename, pose_ref_index, video_format, crop_size, c
     
     
     frames, n, time_series = align_mouse(path_to_file, filename, video_format, crop_size, pose_list, pose_ref_index,
-                                         confidence, pose_flip_ref, bg, frame_count, use_video)
+                                         confidence, pose_flip_ref, bg, frame_count, use_video=use_video)
     
     if check_video and not save:
         play_aligned_video(frames, n, frame_count, path_to_file)
     elif check_video and save:
-        play_aligned_video(frames, n, frame_count, path_to_file, filename=filename, crop_size=crop_size, save=True)
+        play_aligned_video(frames, n, frame_count, path_to_file, save=True, filename=filename, crop_size=crop_size)
 
     return time_series, frames
 
@@ -313,22 +401,24 @@ def alignment(path_to_file, filename, pose_ref_index, video_format, crop_size, c
 def egocentric_alignment(config, pose_ref_index=[0,5], crop_size=(300,300), use_video=False, video_format='.mp4', check_video=False, save=False):
     """ Happy aligning """
     #config parameters
+    if save and not use_video:
+        raise ValueError("If save=True use_video must also be True.")
     config_file = Path(config).resolve()
     cfg = read_config(config_file)
     
     path_to_file = cfg['project_path']
     filename = cfg['video_sets']
-    confidence = cfg['pose_confidence']
     video_format=video_format
     crop_size=crop_size
     
     # call function and save into your VAME data folder
     for file in filename:
-        print("Aligning data %s, Pose confidence value: %.2f" %(file, confidence))
-        egocentric_time_series, frames = alignment(path_to_file, file, pose_ref_index, video_format, crop_size, 
-                                                   confidence, use_video=use_video, check_video=check_video, save=save)
-        np.save(os.path.join(path_to_file,'data',file,file+'-PE-seq.npy'), egocentric_time_series)
-#        np.save(os.path.join(path_to_file,'data/',file,"",file+'-PE-seq.npy', egocentric_time_series))
-        
-    print("Your data is now ine right format and you can call vame.create_trainset()")
+        if not os.path.exists(os.path.join(path_to_file,'data',file,file+'-PE-seq.npy')):
+            print("Aligning data %s, Pose confidence value: %.2f" %(file, confidence))
+            egocentric_time_series, frames = alignment(config, file, pose_ref_index, video_format, crop_size, 
+                                                       use_video=use_video, check_video=check_video, save=save)
+            np.save(os.path.join(path_to_file,'data',file,file+'-PE-seq.npy'), egocentric_time_series)
+        else:
+            print("Aligned video data already exists. To realign this video delete the PE-seq.npy file from the data directory and run this function again")
+    print("Your data is now egocentrically aligned and you can call vame.create_trainset()")
